backend/requirements.txt
anyio==3.6.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.84.0
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
idna==3.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
